修改

+++ src/main/java/com/conan/console/server/entity/master/CostRecord.java
@@ -55,7 +55,7 @@ public class CostRecord {
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    private Float cost_gold;
+    private Double cost_gold;
 
     /**
      *
@@ -64,7 +64,7 @@ public class CostRecord {
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    private Float remain_gold;
+    private Double remain_gold;
 
     /**
      *
@@ -203,7 +203,7 @@ public class CostRecord {
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public Float getCost_gold() {
+    public Double getCost_gold() {
         return cost_gold;
     }
 
@@ -215,7 +215,7 @@ public class CostRecord {
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public void setCost_gold(Float cost_gold) {
+    public void setCost_gold(Double cost_gold) {
         this.cost_gold = cost_gold;
     }
 
@@ -227,7 +227,7 @@ public class CostRecord {
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public Float getRemain_gold() {
+    public Double getRemain_gold() {
         return remain_gold;
     }
 
@@ -239,7 +239,7 @@ public class CostRecord {
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public void setRemain_gold(Float remain_gold) {
+    public void setRemain_gold(Double remain_gold) {
         this.remain_gold = remain_gold;
     }
