+++ src/main/java/com/conan/console/server/entity/master/QueryRecharge.java
@@ -55,7 +55,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    private Float rmb_amount;
+    private Double rmb_amount;
 
     /**
      *
@@ -64,7 +64,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    private Float gold_amount;
+    private Double gold_amount;
 
     /**
      *
@@ -73,7 +73,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    private Float gold_coupon;
+    private Double gold_coupon;
 
     /**
      *
@@ -82,7 +82,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    private Float gold_total;
+    private Double gold_total;
 
     /**
      *
@@ -261,7 +261,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public Float getRmb_amount() {
+    public Double getRmb_amount() {
         return rmb_amount;
     }
 
@@ -273,7 +273,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public void setRmb_amount(Float rmb_amount) {
+    public void setRmb_amount(Double rmb_amount) {
         this.rmb_amount = rmb_amount;
     }
 
@@ -285,7 +285,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public Float getGold_amount() {
+    public Double getGold_amount() {
         return gold_amount;
     }
 
@@ -297,7 +297,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public void setGold_amount(Float gold_amount) {
+    public void setGold_amount(Double gold_amount) {
         this.gold_amount = gold_amount;
     }
 
@@ -309,7 +309,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public Float getGold_coupon() {
+    public Double getGold_coupon() {
         return gold_coupon;
     }
 
@@ -321,7 +321,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public void setGold_coupon(Float gold_coupon) {
+    public void setGold_coupon(Double gold_coupon) {
         this.gold_coupon = gold_coupon;
     }
 
@@ -333,7 +333,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public Float getGold_total() {
+    public Double getGold_total() {
         return gold_total;
     }
 
@@ -345,7 +345,7 @@ public class QueryRecharge {
      *
      * @mbg.generated Tue May 08 00:49:28 CST 2018
      */
-    public void setGold_total(Float gold_total) {
+    public void setGold_total(Double gold_total) {
         this.gold_total = gold_total;
     }
