+++ src/main/java/com/conan/console/server/entity/master/UserRemain.java
@@ -20,7 +20,7 @@ public class UserRemain{
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    private Float gold_amount;
+    private Double gold_amount;
 
     /**
      *
@@ -29,7 +29,7 @@ public class UserRemain{
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    private Float gold_coupon;
+    private Double gold_coupon;
 
     /**
      *
@@ -90,7 +90,7 @@ public class UserRemain{
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public Float getGold_amount() {
+    public Double getGold_amount() {
         return gold_amount;
     }
 
@@ -102,7 +102,7 @@ public class UserRemain{
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public void setGold_amount(Float gold_amount) {
+    public void setGold_amount(Double gold_amount) {
         this.gold_amount = gold_amount;
     }
 
@@ -114,7 +114,7 @@ public class UserRemain{
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public Float getGold_coupon() {
+    public Double getGold_coupon() {
         return gold_coupon;
     }
 
@@ -126,7 +126,7 @@ public class UserRemain{
      *
      * @mbg.generated Mon Apr 23 00:59:28 CST 2018
      */
-    public void setGold_coupon(Float gold_coupon) {
+    public void setGold_coupon(Double gold_coupon) {
         this.gold_coupon = gold_coupon;
     }
