+++ src/main/java/com/conan/console/server/service/CacheService.java
@@ -14,45 +14,45 @@ import com.conan.console.server.utils.ConanUtils;
 public class CacheService {
 
 	@Cacheable(cacheNames = "detailScore")
-	public JSONObject randomList5(String accountName, float score) {
-		List<Float> floatList = new ArrayList<>();
+	public JSONObject randomList5(String accountName, double score) {
+		List<Double> floatList = new ArrayList<>();
 
-		floatList.add(3.3f);// register_info_score0
-		floatList.add(3.3f);// identify_info_score1
-		floatList.add(3.4f);// background_info_score2
+		floatList.add(3.3d);// register_info_score0
+		floatList.add(3.3d);// identify_info_score1
+		floatList.add(3.4d);// background_info_score2
 
 		if (score >= 80) {// 危险
-			floatList.add(7f);// account_growup_score
-			floatList.add(7f);// trade_frequency_score
-			floatList.add(6f);// like_info_score 1危险,2可疑,3安全
+			floatList.add(7d);// account_growup_score
+			floatList.add(7d);// trade_frequency_score
+			floatList.add(6d);// like_info_score 1危险,2可疑,3安全
 		} else if (score >= 60) {// 可疑
 			score = score -2;
-			floatList.add(7f);// account_growup_score
-			floatList.add(7f);// trade_frequency_score
-			floatList.add(4f);// like_info_score 1危险,2可疑,3安全
+			floatList.add(7d);// account_growup_score
+			floatList.add(7d);// trade_frequency_score
+			floatList.add(4d);// like_info_score 1危险,2可疑,3安全
 		} else {// 安全
 			score = score -4;
-			floatList.add(7f);// account_growup_score
-			floatList.add(7f);// trade_frequency_score
-			floatList.add(2f);// like_info_score 1危险,2可疑,3安全
+			floatList.add(7d);// account_growup_score
+			floatList.add(7d);// trade_frequency_score
+			floatList.add(2d);// like_info_score 1危险,2可疑,3安全
 		}
 
-		floatList.add(10f);// trade_process_score
-		floatList.add(10f);// trade_habit_score
-		floatList.add(10f);// logistics_character_score
+		floatList.add(10d);// trade_process_score
+		floatList.add(10d);// trade_habit_score
+		floatList.add(10d);// logistics_character_score
 
-		floatList.add(8.3f);// activity_score0
-		floatList.add(8.3f);// activity_score1
-		floatList.add(8.4f);// activity_score2
+		floatList.add(8.3d);// activity_score0
+		floatList.add(8.3d);// activity_score1
+		floatList.add(8.4d);// activity_score2
 
-		floatList.add(15f);// account_history_score
+		floatList.add(15d);// account_history_score
 
 		int temp = (int) (100 - score) * 10;
 		while (temp > 0) {
 			Random rand = new Random();
 			int index = rand.nextInt(13);
 			if (floatList.get(index) > 0.5 && index != 5) {
-				floatList.set(index, ConanUtils.fix2(floatList.get(index) - 0.1f));
+				floatList.set(index, ConanUtils.fix2(floatList.get(index) - 0.1d));
 				temp = temp - 1;
 			}
 		}
@@ -95,46 +95,46 @@ public class CacheService {
 	@Cacheable(cacheNames = "saftyScore")
 	public JSONObject randomSaftyScoreList5(String accountName) {
 		Random randScore = new Random();
-		float score = randScore.nextInt(40)+20;
+		double score = randScore.nextInt(40)+20;
 		
-		List<Float> floatList = new ArrayList<>();
+		List<Double> floatList = new ArrayList<>();
 
-		floatList.add(3.3f);// register_info_score0
-		floatList.add(3.3f);// identify_info_score1
-		floatList.add(3.4f);// background_info_score2
+		floatList.add(3.3d);// register_info_score0
+		floatList.add(3.3d);// identify_info_score1
+		floatList.add(3.4d);// background_info_score2
 
 		if (score >= 80) {// 危险
-			floatList.add(7f);// account_growup_score
-			floatList.add(7f);// trade_frequency_score
-			floatList.add(6f);// like_info_score 1危险,2可疑,3安全
+			floatList.add(7d);// account_growup_score
+			floatList.add(7d);// trade_frequency_score
+			floatList.add(6d);// like_info_score 1危险,2可疑,3安全
 		} else if (score >= 60) {// 可疑
 			score = score -2;
-			floatList.add(7f);// account_growup_score
-			floatList.add(7f);// trade_frequency_score
-			floatList.add(4f);// like_info_score 1危险,2可疑,3安全
+			floatList.add(7d);// account_growup_score
+			floatList.add(7d);// trade_frequency_score
+			floatList.add(4d);// like_info_score 1危险,2可疑,3安全
 		} else {// 安全
 			score = score -4;
-			floatList.add(7f);// account_growup_score
-			floatList.add(7f);// trade_frequency_score
-			floatList.add(2f);// like_info_score 1危险,2可疑,3安全
+			floatList.add(7d);// account_growup_score
+			floatList.add(7d);// trade_frequency_score
+			floatList.add(2d);// like_info_score 1危险,2可疑,3安全
 		}
 
-		floatList.add(10f);// trade_process_score
-		floatList.add(10f);// trade_habit_score
-		floatList.add(10f);// logistics_character_score
+		floatList.add(10d);// trade_process_score
+		floatList.add(10d);// trade_habit_score
+		floatList.add(10d);// logistics_character_score
 
-		floatList.add(8.3f);// activity_score0
-		floatList.add(8.3f);// activity_score1
-		floatList.add(8.4f);// activity_score2
+		floatList.add(8.3d);// activity_score0
+		floatList.add(8.3d);// activity_score1
+		floatList.add(8.4d);// activity_score2
 
-		floatList.add(15f);// account_history_score
+		floatList.add(15d);// account_history_score
 
 		int temp = (int) (100 - score) * 10;
 		while (temp > 0) {
 			Random rand = new Random();
 			int index = rand.nextInt(13);
 			if (floatList.get(index) > 0.5 && index != 5) {
-				floatList.set(index, ConanUtils.fix2(floatList.get(index) - 0.1f));
+				floatList.set(index, ConanUtils.fix2(floatList.get(index) - 0.1d));
 				temp = temp - 1;
 			}
 		}
